remote_Job/requirements.txt
asgiref==3.3.2
dj-database-url==0.5.0
# Django==3.2.20
django-ckeditor
django-js-asset==1.2.2
django-taggit
django-user-visit==0.4
django-crispy_forms==2.0
pytz==2019.3
sqlparse==0.4.4
ua-parser==0.10.0
user-agents==2.2.0
phonenumbers
django-phonenumber_field
pillow
django-jazzmin